add CNM direct send config
