+++ pkg/process/net/common_windows.go
@@ -1,0 +1,12 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2025-present Datadog, Inc.
+
+package net
+
+import "net/http"
+
+func GetNetworkID(_ *http.Client) (string, error) {
+	return "", nil
+}
